fix import

+++ pandapower/timeseries/data_sources/frame_data.py
@@ -2,7 +2,7 @@
 
 # Copyright (c) 2016-2023 by University of Kassel and Fraunhofer Institute for Energy Economics
 # and Energy System Technology (IEE), Kassel. All rights reserved.
-import numpy as np
+from numpy import int64
 from pandapower.timeseries.data_source import DataSource
 
 try:
@@ -34,8 +34,8 @@ class DFData(DataSource):
         self.df = df
         if multi:
             # casting column and index to int for multi- columns accessing
-            self.df.index = self.df.index.astype(np.int64)
-            self.df.columns = self.df.columns.astype(np.int64)
+            self.df.index = self.df.index.astype(int64)
+            self.df.columns = self.df.columns.astype(int64)
 
     def __repr__(self):
         s = "%s with %d rows and %d columns" % (
